Extract perf tests from unit.
